Add for-each loops
Using the syntax

'for' Ident [, Ident]* '<-' Expr 'do'?

a for-each loop iterating over the expression after the '<-' (read as
element-of) can be created. Note that multiple identities can be
declared, each one effectively acting as a variable definition.

For example

	src : array[i32, 1000] = some_data()

	dst : array[i32, 1000]

	for element, index <- src
		get(dst, index) = element

could be used to copy the array src to dst.