Remove possibility of tuple return values
The definition or assignment of multiple variables separated by ',' is
no longer valid. This simplifies syntax, and wasn't really in line with
the current language design anyways.
The only place where the definition of multiple comma-separated
variables is still allowed is in a for-each loop, to allow access to the
iterator's index.