Reworked the way templates work
Templates are now no longer created using the '[?T]' bounds syntax, but
are instead functions that return a type. An array of 5 integers would
previously have been declared as

	five_ints : Array[i32, 5]

but is now

	five_ints : Array(i32, 5)

which considerably simplifies the syntax. Also note that constaints on
templated types can no longer be expressed directly in the definition,
i.e.

	Array[?T : Copyable] :: // snap

but must instead be explicitly checked in the template function, which
may not actually be an optimal solution.
One option to address this might be to introduce an option 'where'
block, such as

Array :: proc(t : type) -> type where implements(t, Copyable) { /* snap */ }

However, such syntax is not currently supported.