Add alloc_and_parse function
alloc_and_parse combines the common pattern of allocating an ast node
and then parsing into it. This not only shortens the parsing code
significantly, but also makes it more robust, as there can be no
mismatch between the allocated node, and the one passed to parse when
using alloc_and_parse.