Add alloc_and_parse override for vecs
The common pattern of vec.push_back({}) followed by parse(vec.last()) is
now also supported by alloc_and_parse.
This increases the robustness of code using this pattern, as it avoids
mismatches between the push_back and the parse, as well as wrong error
return codes (i.e. a return of false instead of error_out_of_memory in
case push_back fails. This was previously the case for switch cases).