Rework ast::FileModule to contain a vec<Expr>
Previously, FileModule contained a Definition with type
ast::Expr::Module and an ast::Expr::Block containing the entire file's
contents. This was originally intended to make handling of files as
module more straightforward, but this is not currently necessary or
likely to be used in future.
This new method also simplifies the parsing of Blocks, as the previous
special case of the pseudo-Block around the whole file no longer exists.