Rework RobinHoodMap to work with non-string entries
This is necessary to support the updated interface of InputFileSet,
which takes FileId structs instead of path strings, as a given file may
not have a single unique path. Instead of using the path, FileId uses
the Win32 BY_HANDLE_FILE_INFORMATION nFileIndexHigh, nFileIndexLow and
dwVolumeSerialNumber to uniquely identify files after opening them.