Avoid map wrap-around in ThreadsafeMap2.insert
This is necessary as it was previously possible for the loop locking map
cachelines to wrap around the entire map, thus trying to re-lock the first
cacheline and causing a deadlock. To prevent this, a configurable cutoff on the
length of this loop, max_insertion_distance, has been added to
ThreadsafeMap2::InitInfo.