minos: Add `path_remove_[file|directory]` and split async I/O into own `file_[read|write]_async` functions
The latter is mainly necessary because otherwise the `overlapped` parameter is
only used for its offset in case of synchronous operation, while
`out_bytes_read` is left unused in case of asynchronous operation.
Since it should always be clear whether a synchronous or an asynchronous file
is being used, splitting functionality between them does not diminish the API's
ease-of-use.