type_pool: Make `TypeId` an `enum class` instead of a single-member struct
This makes debugging easier, as the Microsoft VS debugger console cannot
execute functions with structures as arguments.
While this implements equality when it was previously purposely left out, this
does not necessary seem like a bad thing, as equality is already used all over
the place by accessing `TypeId.rep`.