interpreter: Fix `MemberInit` initialization to correctly set `type.pending`
This was erroneously being set to the value if there was no type, causing
`force_member_type` to fail, as the value would usually not be of type `Type`,
making its evaluation to a type fail.
Just setting `type.pending` to `AstNodeId::INVALID` if there is no explicit
type fixes the issue.