container: Fix `ReservedVec.reset` to be short-circuit on unchanged commit
Previously the current commit was compared to the commit limit with `<` instead
of `<=`. This meant that system calls were made to achieve exactly no effect in
case of an exact match.
This was especially significant as exact equality is a common case (consider
allocating and the resetting in a loop, never going over the reset limit).
The fix is to simply replace `<` with `<=`.