type_pool: Add `type_discard` to discard a previously created `TypeId`
This allows composite types created for block scopes to be discarded once the
block is exited. This is possible as it is known that the type cannot be
referenced from anywhere else - in particular from another type's
`holotype_id` - as it is not accessible to user code and never compared to
another type or stored otherwise past the block's "lifetime".