interpreter: Add result conversion to `evaluate(AstTag::OpAdd)`
Previously if both summands were `CompInteger`s or `CompFloat`s, but the
`EvalSpec`'s expected type was unequal, but could be reached via implicit
conversion, `evaluate would fail`.
This would e.g. happen with `let x: u32 = 1 + 1`, where the `1`s are
`CompInteger`s, but the result of the `OpAdd` is cast to `u32`.
The reason for this was that the summands' unified type was used in one place
(the actual addition), and the spec's type in another (the result creation via
`fill_spec`). To fix this discrepancy, the code performing addition has had
additional checks added in case of summands of `CompInteger` or `CompFloat`
type, performing implicit conversion as necessary.