ast_pool: Add lowering for nodes with `where` and `OpSet*` nodes
This lowers `AstTag::If`, `AstTag::For`, `AstTag::ForEach` and `AstTag::Switch`
nodes with a `where` clause, by synthesizing a block around them which receives
the `where`'s definitions, and lower `AstTag::OpSet*` to a block containing a
temporary definition, an `AstTag::OpSet`, and the operation of corresponding to
the original `AstTag::OpSet*`.
For the exact resulting ASTs see the comments on `lower_expr_with_where` and
`lower_set_op` in core/ast_pool.cpp.
In its current implementation this requires an extra `memcpy` of the entire
AST, which is could be fixed by integrating it into
`copy_preorder_to_postorder`, which would however make it even more fiddly, but
could still be done at a later stage.