type_pool: Remove `IncompleteMemberIterator` and associated functions
`IncompleteMemberIterator` was inherently problematic, as an incomplete member
selected by `incomplete_members_of` or `next` might be completed before the
next call to `has_next` / `next`, making the iterator return a complete member
against its spec.
The simpler and more reliable alternative is to use `MemberIterator` and check
whether the iterated members are complete at the call-site of `next`.