common: Redefine `ASSERT_UNREACHABLE` to use compiler-specific builtins instead of divide-by-zero
The intent behind the divide-by-zero was to let the compiler optimize around
the ensuing Undefined Behaviour, effectively making it super-easy to prove that
the relevant code / branch could not be reached.
The problem is however that msvc instead - very helpfully - generates error
C2124 "divide or mod by zero" for the old version.
Using `__builtin_unreachable()` on gcc / clang, and __assume(false) on msvc
should have the desired effect, but actually work.