interpreter: Unconditionally initialize `NumericType` in `evaluate(AstTag::OpShift[L|R])` that was causing spurious warnings from gcc
Due to the function's exact control flow, the affected variable was only
initialized and used when `lhs`'s was of `TypeTag::Integer`.
However, this does not seem to be evident to gcc's optimizer, causing a false
positive warning regarding use of a possibly uninitialized variable in
optimized builds.
To avoid this, the simplest option appears to be to default-initialize the
affected variable. In terms of runtime costs, this should be basically free.