Rework error handling to no longer use `setjmp`/`longjmp`
This major overhaul was necessary because the implementation of `longjmp` on
windows was extremely unreliable for some reason (possibly related to
`noexcept`?).
Note that `setjmp` / `longjmp` is still used locally in `parse.cpp`, but this
more constrained usage appears to work just fine for now.
This commit additionally includes a rework of `OptPtr<T>` to `Maybe<T>`,
making it more general (allowing anything with a self-typed static
`INVALID` member instead of just pointers) and extended result formatting /
filtering in make-all.ps1.