Create new interpreter that runs on bytecode instead of AST
This is a major rewrite of a lot of the codebase, meaning there are certainly
still oh-so-many regressions to be crushed.
Note also that there are still a few files from the old version left over
(core/interpreter.cpp, core/global_value_pool.cpp, ...). These will be removed
in upcoming commits.