main: Add "Failure" output upon error exit for clarity
This avoids the problem of not instantly knowing whether the process ran to
completion and failed in a controlled way or simply crashed when there are no
errors in the log. This should of course never happen, but there have been a
few bugs that led to exactly this situation already.